+++ mtx_transform.h
@@ -6,7 +6,7 @@
 /*   By: zcanales <zcanales@student.42urduliz.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/20 11:12:36 by zcanales          #+#    #+#             */
-/*   Updated: 2022/04/20 12:13:06 by zcanales         ###   ########.fr       */
+/*   Updated: 2022/04/20 13:48:56 by zcanales         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,12 +16,12 @@
 //# include "matrix.h" -> Añadir
 
 //mtx_transform.c
-float	**translation_mtx(float x, float y, float z);
-float	**scaling_mtx(float x, float y, float z);
-float	**shearing_mtx(float shear_arr[6]);
+double	**translation_mtx(double x, double y, double z);
+double	**scaling_mtx(double x, double y, double z);
+double	**shearing_mtx(double shear_arr[6]);
 //mtx_rotations.c
-float	**rotatation_x_mtx(float grade);
-float	**rotatation_y_mtx(float grade);
-float	**rotatation_z_mtx(float grade);
+double	**rotatation_x_mtx(double grade);
+double	**rotatation_y_mtx(double grade);
+double	**rotatation_z_mtx(double grade);
 
 #endif
